add extra output to trace an intermittent problem
Rarely when pscreen is used for the first time after systerm restart it
will error out on the lines changed for an unknown reason.  I can only
suppose that it is related to dead screen sessions, but this should
already be taken care of.  The extra output should help explain what the
reason for the error is.

Traceback (most recent call last):
  File "/root/.local/bin/pscreen", line 133, in <module>
    main()
  File "/root/.local/bin/pscreen", line 129, in main
    Profiles(**parse_args(prof_dir))
  File "/root/.local/bin/pscreen", line 28, in __init__
    self.find_sessions()
  File "/root/.local/bin/pscreen", line 63, in find_sessions
    if not match.group('name') in self.profile_dict:
AttributeError: 'NoneType' object has no attribute 'group'